pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "snowexsql"
dynamic = ["version"]
description = "SQL Database software for SnowEx data"
keywords = ["snowex", "sql", "database", "snow"]
readme = "README.rst"
requires-python = ">=3.8"
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10'
]
dependencies = [
    "utm>=0.5.0,<1.0",
    "geoalchemy2>=0.6,<1.0",
    "geopandas>=0.7,<2.0",
    "psycopg2-binary>=2.9.0,<2.10.0",
    "rasterio>=1.1.5",
    "SQLAlchemy >= 2.0.0",
]

[project.optional-dependencies]
dev = [
    "factory_boy<4",
    "pytest-factoryboy<3",
    "pytest",
    "pytest-cov",
    "sphinx-autobuild<=2024.5",
]
docs = [
    "ipython>7.0,<9.0",
    "nbconvert>=6.4.3,<6.5.0",
    "nbsphinx==0.9.4",
    "pandoc==1.0.2",
    "plotly==5.22.0",
    "sphinx-gallery==0.9.0",
    "sphinx>=7.1,<7.4",
    "sphinxcontrib-apidoc==0.3.0",
    "jupyter-book>=1.0.2,<1.1",
    "pyyaml<6.1"
]
all = ["snowexsql[dev,docs]"]

[project.license]
file = "LICENSE"

[project.urls]
Homepage = "https://github.com/SnowEx/snowexsql"
Documentation = "https://snowexsql.readthedocs.io"
Repository = "https://github.com/SnowEx/snowexsql.git"
Issues = "https://github.com/SnowEx/snowexsql/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "snowexsql/_version.py"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.build.targets.sdist]
exclude = ["/tests"]
